Move getNonce
