+++ src/helpers.ts
@@ -32,3 +32,13 @@ export async function loadBrokerProfiles(): Promise<MqttBrokerConfig[] | undefin
     let brokerProfiles = await config.get<Array<MqttBrokerConfig>>("brokerProfiles");
     return brokerProfiles;
 }
+
+export function getNonce() {
+    let text = "";
+    const possible =
+      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+    for (let i = 0; i < 32; i++) {
+      text += possible.charAt(Math.floor(Math.random() * possible.length));
+    }
+    return text;
+}
